RecorTimer.py: disable instanceTrashcan markDirty
disabled.
reproduce error:
we start a timer recording:
MultiSelectionEPG (current event) press the green button and add the timer, if for any circumstance we want to quickly eliminate said timer (in seconds) either because we have made a mistake or for any circumstance, the system literally hangs (system block)

2.- We reboot the system and E2 insists on recording.blem persists since this instance calls the trash constantly.
this code wants to mark dirty to the trash and the recording module persists in making said recording, as a consequence of all this it does not start the system even after restarting.