[Pixmap.py] Stop CHANGED_CLEAR clearing the pixmap inappropriately
NOTE: This corrects an earlier (March 2018) fix to Pixmap.py

The instance Pixmap should only be cleared on CHANGE_CLEAR if the Pixmap is being taken from the renderer source's Pixmap attribute (when it has one).

This fixes problems seen in the display of the tuner type icon in the InfoBar in the iSkin-* and Q-FHD_BLACK skins, where the icon would be displayed after a reboot (or GUI restart), but disappear after the first channel change.

This fix was created by Prl.  Thanks to Prl for alerting us to the issue and its resolution.