PliExtraInfo.py: refactor code
 [PliExtraInfo] Fix up logic on "not feraw" return

In PliExtraInfo.getText(), move conversions that don't use feraw,
or return non-empty results on "not feraw" to before the empty
string return on "not feraw" ("PIDInfo", "ServiceRef", "TransponderInfo").
 [PliExtraInfo] Make gamma conversion more robust

In PliExtraInfo.createResolution() make the conversion of the
gamma/dynamic range to a string more robust to out-of-range values.

    develop  [PliExtraInfo] Refactor PliExtraInfo to make it more flexible

Refactor PliExtraInfo so that conversions that produce multiple
conversion strings (like "All" and "TransponderInfo") are generated
from tables rather than directly in code.

Make use of this facility to add a "User" conversion where the skin
author may customise which conversions to include.

Also add a "TransponderFrequencyMHz" conversion to display the transponder in
MHz to 3 decimal places.

These facilities will be used in changes to the pre-loaded skins
that allow for more than 4 tuner state indicators to be shown in
the infobar.

    develop

fix bug thx huevos

Refactor code thx prl001