[EpgList] Reworked string (#3117)
The "prefix" can be either "" (empty string) or "+", thus by removing it from the translations we fix the following warning when updating the .po files:

'msgid' format string with unnamed arguments cannot be properly localized:
The translator cannot reorder the arguments.
Please consider using a format string with named arguments,
and a mapping instead of a tuple for the arguments.

In addition, we have one less string to translate.
The new string is already used in enigma2, so no update needed in the translations.