[SoftwareManager] Reworked 2 text strings (#3127)
This is one phrase, so it should be translated as one string.
The existing string parts do not translate the same when combined into one sentence. All existing translations are NOT updated to be done again.

Use ngettext() to offer singular and plural forms of string.
Existing translations will be deleted!