SkinSelector.py: added descriptions in display skin choice
with this change we have independent descriptions for display skin.
Before we had the same description for all of them, we must bear in mind that the skin displays do not have resolution, therefore we cannot involve them in the skin descriptions with resolution variables since it returns None and this is what we see in the description.
now if we can see in the description that we are selecting display skin as a name.
I add directory delete default:
If we put by default it is what we see in the UI, nevertheless we can create some directories in the display path together with skin_default, these directories that we can call as we want, we can enter the XML that we want designed and thus be able to select it and choose it ... before We couldn't see those directories.