Timeshift.py: use simpler code for footnote
In boxes with driver problems to detect this function, all the strings formats for the debugging format may not be met.
Played on sf8008 box with NAS network device mounted.
with this we ensure a simpler debugging and without the possibility that this causes a crash.
note that we already have a footnote description + a TYPE ERROR that makes the statement ... this is enough.