EpgSelection.py: classify mapto "blue" with or without EPGSearch plugin installed
With this change we assign a new data structure in the dictionary, without mapto blue for this action with the EPGSearch plugin installed.

EPGSearch in your action with blue button should assign Options List or Search in in channel selection screen, if the action for the blue button remains in the dictionary with EPGSearch installed, they are not executed for this plugin, take control Simple EPG with your "Select Channel" action to search SimpleChannelSelection for a channel for your Simple EPG.

Another error added to the above mentioned is that we will see the text of the EPGSearch "Search" or "Options List" action on its blue button and not the "Select Channel" action but we actually execute it if we press blue "Select Channel".

This sorts the dictionary with or without EPGSearch installed by assigning the correct actions on its blue button for EPGSearch and EPG SINGLE.