CutListEditor: fix edit  broken for python2
C++ for CutLisEditor implements PyLong_Check and PyInt_Check
by default we had invalid integer int declaration for python2, use long declaration for python2 and int for python3.
it is written conditionally assigning two types of declarations depending on the type of python used, this fixes broken CutListEditor in python2.