ImportChannels - ChannelsImporter: add alternative ChannelsImporter module
Standard HTTP 1.1 used by Request is causing problems in programming with Python3, the result is not having EPG and channels available from the ImportChannels module, this makes a call as an alternative resource to solve this situation.

Description is added specifying:
If the ClientMode Module is used with ChannelsImporter, if we want to do an EPG download from the server, the epg.dat file must be in internal flash on this server, not fixed in Python3.

ChannelsImporter:
- Imports for Python3 are adapted.
- fix TypeError: 'ConnectionRefusedError' object is not subscriptable