ImportChannels - Client Mode : unlink the import of these two modules
Although the parameters are not active, the remote tuner setup will take values present in Client Mode.
in this way we unlink the actions from the parameters specifying which module is active.
the ERROR notification "do not set manual IP" is set with an ID of ERROR.
Exceptions are added to mitigate possible file copy problems.