Timeshift: Create Timeshift directory on reboot
We reduce code, to do the same function.
this code executes the creation of a user-allocated timeshift directory ( backup restore ) if it does not exist.

With this, a directory is always assigned for timeshift, either the default directory or a directory previously created by the user in copy restoration, the system remembers it and creates it again if it does not exist.

If the timeshift directory is now different from the restore backup directory, pathStatus assigns it as a flash directory -> Timeshift cannot be executed.