UsageConfig.py: config.usage.timeshift_path for create directories
for create Timeshift directories:

With the new code, there is no need to start the code to create specific mount directories.
The new code assigns to config.timeshift.path.value with the makedirs method this task is in charge of creating the directories without the need to apply the creation of directories in the mount points.
For USB the new written variable will do it for this type of mounts.
config.usage.timeshift_path.value Will create directories for DefaultValue and those assigned by users.
This cleans up the code even more.