[skin.py] Support new ConfigList and windowstyle fonts
This change adds support for the new screen attributes "entryFont" and "valueFont".  These fonts are used for the "entry" and "value" elements, respectively, of each line in a ConfigList display.  Previously the generic "font" attribute was used for the "entry" string and "secondfont" was used for the value string.  These variables were not obvious and, in particular, "secondfont" rarely used.  The new names better relate to what these values affect.  The old names if used will still work but a deprecation notice will be logged.

The "windowstyle" skin configuration block now has a new tag "configList" that has two attributes "entryFont" and "valueFont" to specify the default fonts to be used for all ConfigList based screens.  For example:
<configList entryFont="Regular;25" valueFont="Regular;25" />