TimerEditList: set defaults to not display duplicate descriptions
the description in the TimerEntry's ConfigList must be passed via getExtendedDescription(), with getShortDescription() and replace sends empty.
thanks to @ims21   code with replaceEPGSeparator(code) with config "space": " " + "nothing" will fix this problem.
so by default to not see duplicate event description in TimerEditList this setting:
- config.epg.fulldescription_separator=nothing
- config.epg.replace_newlines=space

It is not easy to grasp it since in the Rytec sources there are different ways to receive the EPG information.