Language: add missing import and some corrections
It is eliminated to translate phrase when we only have "English" as the only available language.
config.pluginbrowser.languages_po view PO sources + language packs, if you enable view language and for PO sources.
LanguageSelection is set to True if I use the yellow button install new language, save is set to config.pluginbrowser.languages_po.value for  if yellow button is used.

It is not allowed to remove language from the browser, it can only be executed in the LanguageSelectiton Module.

Revised some descriptions to establish the correct action of each function.