[skin.py] Provide a default font alias height
This change provides a default value for font alias heights if the height attribute is not specified in the skin.  In the past this was silently ignored.  When the parseInteger() method was added this situation was detected and the issue reported and the new parse code then substituted a safe value.  This code change adds a default value to be used if the attribute is undefined in the skin.  Doing this eliminates the "[Skin] Error: The value 'None' is not a valid integer!" error in the logs.