RecordTimer.py: fix integer for timer repetead and fix bug begTime
Fix this bug:
TypeError: in method 'eEPGCache_lookupEventTime', argument 3 of type 'time_t'

C++ time_t in queryTime already starts begTime, the time sequence to start the event time.

Fix this problem:

TimerEntry does not detect booleans for repeating timers, they are integers to sort for repeating weekly, daily, Mon-Fry, or user-defined.

The consequence is that if we restart the UI, it will lose the assigned data, being assigned to an else in TimerEntry ->Weekly.

here it was changed to boolean:

https://github.com/norhap/enigma2-openvision-1/commit/fe5f91277a1b3b82e5202a540af34147d7933d25