[MenuList.py][ConfigList.py][List.py][ChoiceList.py] update to code
MenuList.py:

- Use the new navigation methods in the C++ code.
- Keep the old navigation methods as pointers to the new methods until the clean up is completed.
- Small code move to make the code more readable.

ChoiceList.py:

- Remove hard coded enableWrapAround
Now that the skin "windowstyle" allows for the skin's default of enableWrapAround to be set it is no longer necessary to hard code the default.  The system default has been adjusted to set the system default to enabled.  The skin default can be set in the "listbox" tag in the "windowstyle" block of the skin.  As before individual screens can still have an "enableWrapAround " attribute to modify the settings for any individual screen.

ConfigList.py:

- Use the new navigation methods in the C++ code.
- Keep the old navigation methods as pointers to the new methods until the clean up is completed.
- Improve some variable names to eliminate PEP8 errors.
- Small code move to make the code more readable.

List.py:

- Use the new navigation methods in the C++ code.
- Keep the old navigation methods as pointers to the new methods until the clean up is completed.
- Improve some variable names to eliminate PEP8 errors.

Co-Authored-By: IanSav <IanSav@users.noreply.github.com>