[services] Add proper return values in start procedures
to properly handle problems e.g. with the tuner or streaming

[Navigation] Streaming: Retry play service
If a stream is active which blocks a tuner on the same box,
changing the channel might not work. The stream client is stopped
before zapping, but the streamserver still blocks the tuner when e2
tries to allocate the tuner.
Use an eTimer to give the streamserver the time to release the tuner
and then try again.

Co-Authored-By: Betacentauri <betacentauri@users.noreply.github.com>