[skin.py] Add a new "wrap" attribute
This change adds a new widget attribute "wrap" to be the opposite of the "noWrap" attribute.  This avoids the double negative of setting no noWrap.  This change is comparable to the recent change of "selection" being created as the opposite of "selectionDisabled".

The deprecation warning on "selectionDisabled" has been removed. The function of "selection" and "selectionDisabled" has been updated, like the new "wrap" and "noWrap", to be fully symmetrical.