InputDevice - keyboard- Rc Add button remapping facility
This code uses a companion component in the C++ RCinput code to allow specific remote controls to have buttons remapped.

The rcpositions.xml file for a remote control can have a "keyid=" attribute that nominates the key id transmitted by the remote control and a "remap=" attribute that nominates the key id value that should be send to Enigma2.  The original keyid value is no longer seen by Enigma2.  It is if the remote control hardware/drivers have been changed to emit the remapped code.  For example, the gb4 remote control for the GigaBlue UHD UE 4K could have the following line in rcpositions.xml:
    <button keyid="KEY_F1" remap="KEY_HELP" name="HELP" title="HELP" pos="26,409" shape="rect" size="28,12" />
This takes the KEY_F1 code sent by the remote control and sends a KEY_HELP code to Enigma2.  That is, this GigaBlue remote control now has a HELP button for the loss of the F1 button.

InputDeviceSetup Refactor:
* init Setup value in XML InputDevices
* Add class remoteControl, keyboard.
iInputDevices, iRcTypeControl -> inputDevices RemoteControl
- Remove Components Keyboard
ImputDevices.py: Refactor
* Add identifier remote control remotes.xml
* Add Remap change name class for ImputDeviceSetup
* change name def
UsageConfig.py:
- Add KEYIDS for new  keyboard
KeyBindings.py:
- config.plugins.remotecontroltype.rctype -> config.inputDevices.rcType

Co-Authored-By: IanSav <IanSav@users.noreply.github.com>