PowerTimer.py: fix timer type "wakeup"
With this commit we fix the "wake up" timer type in PowerTimer.
If we simultaneously have a "wake up" type timer in PowerTimer and in turn have the "wake up value enabled" in the SleepTimer module, it was not work before.

PowerTimerEntry.py:
Wakeup to standby timers are not enabled unless SleepTimer is not enabled.
If "wakeup to standby" is enabled in SleepTimer, this module is in charge of executing this function.