[DVD] Fix infobar show / hide and introduce angle label text
This fix tries to disable the InfoBar showing in DVD menues which is annoying because it allows actions in the menu only if the InfoBar is hidden.
Also a new text is added to show the camera perspective of DVDs that support it. MetrixHD does not have an Icon for it. So thus the textual description.
Furthermore there is a fix to a print statement for logging / debugging purpose.

Co-Authored-By: Alexander Meinke <4l3x2k@users.noreply.github.com>