Add separate text border attributes
This change adds two new attributes "textBorderColor" and "textBorderWidth" to specify that a border color and its width is to specifically apply to the text within a widget as opposed to the widget itself.  A pending commit will redefine "borderColor" and "borderWidth" to apply to the widget itself.  This is being done to make the skin language more consistent among the different widgets and controls.

[eLabel]
* Add setTextBorderColor and setTextBorderWidth to set the border color and width on text within a widget.

[skin.py]
* Add textBorderColor and textBorderWidth attributes to set the border color and width of text within a widget.