PluginBrowser.py: remove chek internet wit checkInternetAccess
checkInternetAccess = 2 (other error internet )

These other errors could be the crash of the ip-api.com website, it is not a good idea to only have this argument to check the network.
I have recently had NO access to this website due to hardware problems on my LOCAL network, this slowed down the receiver, and you also need to have a more consistent argument to verify the network, not just a website.