MediaPlayer.py: always save a default directory never None!
reproduce:
Enter edit the settings and exit this screen again, it will ask us if we want to exit without saving the configuration.
Enter this screen again but this time save with the GREEN button.
What have we done? save the directory config.mediaplayer.defaultDir with value to None.
This is a serious mistake, if we do this we will enter a dead end loop since filelist will not be able to obtain the mount points, this value cannot be None.

With this change make sure that the value is not none, we take a reference value either the one chosen by the user or the default movie path.

add missing key mapto "save" and define keySave.