VideoMode: Set available modes if there is more than one video port
ActionMap: with the new changes.
Pressing OK or Menu sets the options to choose from in the item menu.
The only save key is the GREEN button.

Before: pressing MENU exited the AUDIO and VIDEO settings menu.
Pressing OK offered the item selection menu.
We follow the dynamics of the interface in all configuration screens.